Navigation pages added, profile page added (blank), camera moved to scan page.
